[Ahead-of-time Subcomponents] Refactor logic relating to modifiable bindings out of ComponentBindingExpressions and into ModifiableBindingExpressions.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=214812986
